add usda.py, move is usda_elderly from snap.py to usda.py
